Renaming some code
