+++ app/src/main/java/co/touchlab/kampkit/android/MainApp.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import android.content.Context
 import android.content.SharedPreferences
 import android.util.Log
-import co.touchlab.kampkit.AndroidBreedViewModel
+import co.touchlab.kampkit.BreedViewModel
 import co.touchlab.kampkit.AppInfo
 import co.touchlab.kampkit.initKoin
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -17,7 +17,7 @@ class MainApp : Application() {
         initKoin(
             module {
                 single<Context> { this@MainApp }
-                viewModel { AndroidBreedViewModel() }
+                viewModel { BreedViewModel() }
                 single<SharedPreferences> {
                     get<Context>().getSharedPreferences("KAMPSTARTER_SETTINGS", Context.MODE_PRIVATE)
                 }
