--- shared/src/androidMain/kotlin/co/touchlab/kampkit/AndroidBreedViewModel.kt
@@ -1,12 +1,0 @@
-package co.touchlab.kampkit
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-
-class AndroidBreedViewModel : ViewModel(),
-    BreedViewModel by BreedViewModelImpl() {
-
-    init {
-        initWithScope(viewModelScope)
-    }
-}
